Moved src/gavicore/ui/providers/panel to stc/gavicore/ui
