+++ gavicore/src/gavicore/ui/panel/registry.py
