+++ gavicore/src/gavicore/ui/panel/widgets/_util.py
