+++ gavicore/src/gavicore/ui/panel/widgets/array.py
