+++ gavicore/src/gavicore/ui/panel/widgets/bbox.py
