+++ gavicore/src/gavicore/ui/panel/widgets/labeled.py
