+++ gavicore/tests/ui/panel/test_field.py
@@ -6,7 +6,7 @@ from unittest import TestCase
 
 from gavicore.models import Schema
 from gavicore.ui import FieldMeta
-from gavicore.ui.providers.panel import PanelField
+from gavicore.ui.panel import PanelField
 
 from .schema2ui import load_schemas
