--- gavicore/tests/ui/providers/panel/__init__.py
@@ -1,3 +1,0 @@
-#  Copyright (c) 2026 by the Eozilla team and contributors
-#  Permissions are hereby granted under the terms of the Apache 2.0 License:
-#  https://opensource.org/license/apache-2-0.
